- Made User registration functional - added editorConfig - updated UrlMappings - updated registerPage.html
