+++ .editorconfig
@@ -1,0 +1,15 @@
+root = true
+
+[*]
+charset = utf-8
+end_of_line = lf
+insert_final_newline = true
+indent_style = space
+indent_size = 4
+trim_trailing_whitespace = true
+
+[*.groovy]
+indent_style = space
+indent_size = 3
+max_line_length = 120
+insert_final_newline = true
