+++ src/main/resources/static/views/loginPage.html
@@ -9,7 +9,7 @@
                 <label>Username</label>
                 <input id="usernameInputField" type="text" placeHolder="Username" required/>
                 <label>Password</label>
-                <input id="passwordInputField" type="text" placeHolder="Password" required/>
+                <input id="passwordInputField" type="password" placeHolder="Password" required/>
 
                 <button onclick="login()" type="submit">Login</button>
             </div>
@@ -18,7 +18,7 @@
                 <span class="psw">Forgot <a href="#">password?</a></span>
             </div>
             <div>
-                <span class="psw">Don't have an account? <a href="/register">Register</a></span>
+                <span class="psw">Don't have an account? <a href="/registerPage">Register</a></span>
             </div>
         </form>
     </div>
